Edge detection works well now

+++ main.py
@@ -10,7 +10,7 @@ def get_args():
     parser = argparse.ArgumentParser()
     parser.add_argument('--file_path', type=str, default='data/trainingsvideo.avi')
     parser.add_argument('--output_path', type=str, default=None)
-    parser.add_argument('--sample_frequency', type=int, default=2)
+    parser.add_argument('--sample_frequency', type=float, default=2)
     args = parser.parse_args()
     return args
